+++ frontend/features/CreateTask/constants.ts
@@ -1,0 +1,23 @@
+import { SubjectName } from "@/types/task/new/subject";
+
+export const priorities = [
+    { value: 1, label: "とても低い" },
+    { value: 2, label: "低い" },
+    { value: 3, label: "普通" },
+    { value: 4, label: "高い" },
+    { value: 5, label: "とても高い" },
+  ];
+
+export const subjectLists: SubjectName[] = [
+    "英語",
+    "数学",
+    "現代文",
+    "古文",
+    "日本史",
+    "世界史",
+    "地理",
+    "物理",
+    "化学",
+    "生物",
+    "地学",
+  ];
